Move BaseStateUnemployment class and StateEmployment interface
