+++ src/Classes/StateUnemployment.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Appleton\Taxes\Classes;
+
+interface StateUnemployment
+{
+    public function getTaxCredit();
+}
