+++ src/Countries/US/Colorado/ColoradoUnemployment/ColoradoUnemployment.php
@@ -2,7 +2,7 @@
 
 namespace Appleton\Taxes\Countries\US\Colorado\ColoradoUnemployment;
 
-use Appleton\Taxes\Countries\US\FederalUnemployment\BaseStateUnemployment;
+use Appleton\Taxes\Classes\BaseStateUnemployment;
 
 abstract class ColoradoUnemployment extends BaseStateUnemployment
 {
