--- src/Countries/US/FederalUnemployment/StateUnemployment.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Appleton\Taxes\Countries\US\FederalUnemployment;
-
-use Appleton\Taxes\Classes\BaseTax;
-
-interface StateUnemployment
-{
-    public function getTaxCredit();
-}
