+++ src/Countries/US/FederalUnemployment/V20170101/FederalUnemployment.php
@@ -3,7 +3,7 @@
 namespace Appleton\Taxes\Countries\US\FederalUnemployment\V20170101;
 
 use Appleton\Taxes\Classes\Payroll;
-use Appleton\Taxes\Countries\US\FederalUnemployment\StateUnemployment;
+use Appleton\Taxes\Classes\StateUnemployment;
 use Appleton\Taxes\Countries\US\FederalUnemployment\FederalUnemployment as BaseFederalUnemployment;
 use Appleton\Taxes\Traits\HasWageBase;
 use Illuminate\Database\Eloquent\Collection;
