+++ src/Countries/US/Georgia/GeorgiaUnemployment/GeorgiaUnemployment.php
@@ -2,7 +2,7 @@
 
 namespace Appleton\Taxes\Countries\US\Georgia\GeorgiaUnemployment;
 
-use Appleton\Taxes\Countries\US\FederalUnemployment\BaseStateUnemployment;
+use Appleton\Taxes\Classes\BaseStateUnemployment;
 
 abstract class GeorgiaUnemployment extends BaseStateUnemployment
 {
