+++ src/Countries/US/NewYork/NewYorkUnemployment/NewYorkUnemployment.php
@@ -2,7 +2,7 @@
 
 namespace Appleton\Taxes\Countries\US\NewYork\NewYorkUnemployment;
 
-use Appleton\Taxes\Countries\US\FederalUnemployment\BaseStateUnemployment;
+use Appleton\Taxes\Classes\BaseStateUnemployment;
 
 abstract class NewYorkUnemployment extends BaseStateUnemployment
 {
