+++ src/Countries/US/Wisconsin/WisconsinUnemployment/WisconsinUnemployment.php
@@ -2,7 +2,7 @@
 
 namespace Appleton\Taxes\Countries\US\Wisconsin\WisconsinUnemployment;
 
-use Appleton\Taxes\Countries\US\FederalUnemployment\BaseStateUnemployment;
+use Appleton\Taxes\Classes\BaseStateUnemployment;
 
 abstract class WisconsinUnemployment extends BaseStateUnemployment
 {
